moving folders and troubleshooting
