--- src/client/features/tasks/NewNote.jsx
@@ -1,26 +1,0 @@
-import { useState } from "react";
-// import { useCreateTaskMutation } from "./taskSlice";
-import { useCreateNoteMutation } from "../../store/Notes";
-
-/** Form for creating new tasks */
-export default function NewNote() {
-  const [note, setNote] = useState("");
-  const [createNote] = useCreateNoteMutation();
-
-  const create = async (evt) => {
-    evt.preventDefault();
-    createNote({ note });
-  };
-
-  return (
-    <form onSubmit={create}>
-      <input
-        type="text"
-        value={note}
-        onChange={(e) => setNote(e.target.value)}
-        required
-      />
-      <button>Create</button>
-    </form>
-  );
-}
